[refactor♻️] PlaceLogService에서  BookmarkService분리 (#58)
