+++ src/main/java/com/spoteditor/backend/placelog/service/PlaceLogBookmarkService.java
@@ -1,0 +1,11 @@
+package com.spoteditor.backend.placelog.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public interface PlaceLogBookmarkService {
+
+    void addBookmark(Long userId, Long placeLogId);
+
+    void removeBookmark(Long userId, Long placeLogId);
+}
