+++ src/main/java/com/spoteditor/backend/placelog/service/PlaceLogService.java
@@ -18,9 +18,5 @@ public interface PlaceLogService {
 
     void publishPlaceLog(Long userId, Long placeLogId);
 
-    void addBookmark(Long userId, Long placeLogId);
-
-    void removeBookmark(Long userId, Long placeLogId);
-
     void removePlaceLog(Long userId, Long placeLogId);
 }
